refactor(release): rename check_status.py script to release_manager.py
