+++ .github/workflows/release-pr.yml
@@ -57,4 +57,4 @@ jobs:
           if [ "$DRY_RUN_INPUT" == "true" ]; then
             EXTRA_ARGS="--dry-run"
           fi
-          ./.agents/skills/a2ui-release-sdks/scripts/check_status.py --create-pr $EXTRA_ARGS
+          ./.agents/skills/a2ui-release-sdks/scripts/release_manager.py --create-pr $EXTRA_ARGS
